refactor: change undefined to option, update import react
