+++ src/components/errorBoundary/index.tsx
@@ -1,9 +1,9 @@
-import React from 'react';
+import { Component, ErrorInfo, ReactNode } from 'react';
 import { ErrorBoundaryProps, ErrorBoundaryState } from '../../interfaces/errorBoundary';
 import styles from './errorBoundary.module.scss';
 import Button from '../button';
 
-export default class ErrorBoundary extends React.Component<ErrorBoundaryProps, ErrorBoundaryState> {
+export default class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
   constructor(props: ErrorBoundaryProps) {
     super(props);
     this.state = {
@@ -19,12 +19,12 @@ export default class ErrorBoundary extends React.Component<ErrorBoundaryProps, E
     window.location.reload();
   };
 
-  componentDidCatch(error: Error, info: React.ErrorInfo): void {
+  componentDidCatch(error: Error, info: ErrorInfo): void {
     this.setState({ hasError: true });
     console.log('Handled error:', error, 'Info:', info);
   }
 
-  render(): string | number | boolean | Iterable<React.ReactNode> | JSX.Element | null | undefined {
+  render(): string | number | boolean | Iterable<ReactNode> | JSX.Element | null | undefined {
     const { hasError } = this.state;
     const { children, fallback } = this.props;
