+++ src/interfaces/store.ts
@@ -1,8 +1,8 @@
 import { Item } from './resultSection';
 
 export interface Data {
-  items: Item[] | undefined;
   page: number;
   limit: number;
   total: number;
+  items?: Item[];
 }
